backend/requirements.txt
# ============================================
# MEDICONNECT BACKEND - REQUIREMENTS
# Python 3.11+ Compatible
# ============================================

# -------------------------------------------
# DJANGO CORE
# -------------------------------------------
Django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0

# -------------------------------------------
# AUTHENTICATION
# -------------------------------------------
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0

# -------------------------------------------
# DATABASE
# -------------------------------------------
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Supabase Storage (for Health Records)
supabase==2.3.4
storage3==0.7.0

# -------------------------------------------
# REDIS & CACHING
# -------------------------------------------
redis==5.2.1
django-redis==5.4.0

# -------------------------------------------
# CELERY (Background Tasks)
# -------------------------------------------
celery==5.4.0
kombu==5.4.2

# -------------------------------------------
# CHANNELS (WebSocket)
# -------------------------------------------
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# -------------------------------------------
# FIREBASE ADMIN
# -------------------------------------------
firebase-admin==6.6.0

# -------------------------------------------
# ML & NLP LIBRARIES
# -------------------------------------------
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2

# spaCy for NLP
spacy==3.8.3

# -------------------------------------------
# AZURE SERVICES
# -------------------------------------------
# azure-cognitiveservices-speech==1.42.0
#Http requests
requests==2.32.3
# openai>=1.0.0
# azure-cognitiveservices-speech>=1.32.0

#Ai service (free)
groq>=0.4.0

#Translation (free)
deep-translator>=1.11.4

# -------------------------------------------
# APScheduler (Task Scheduling) Lightweight and more robust than raw daemon threads
# -------------------------------------------
apscheduler==3.10.4

# -------------------------------------------
# FILE HANDLING
# -------------------------------------------
Pillow==11.0.0
python-magic-bin==0.4.14

# -------------------------------------------
# PDF GENERATION
# -------------------------------------------
reportlab==4.2.5

# -------------------------------------------
# ENVIRONMENT & CONFIGURATION
# -------------------------------------------
python-dotenv==1.0.1
python-decouple==3.8

# -------------------------------------------
# UTILITIES
# -------------------------------------------
pytz==2024.2
python-dateutil==2.9.0.post0

# -------------------------------------------
# API DOCUMENTATION (Optional)
# -------------------------------------------
drf-yasg==1.21.8

# -------------------------------------------
# DEVELOPMENT & TESTING
# -------------------------------------------
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
coverage==7.6.8

# -------------------------------------------
# CODE QUALITY
# -------------------------------------------
flake8==7.1.1
black==24.10.0

# -------------------------------------------
# PRODUCTION
# -------------------------------------------
gunicorn==23.0.0
whitenoise==6.8.2